backend/requirements.txt
flask==2.3.3
python-dotenv==1.0.0
google-generativeai==0.3.1
flask-cors==4.0.0
